!4: Implement graph processing library
Implements a basic (non distributed) graph processing library. It is made in a way such that it can be made distributed in the future and already supports being run in parallel (it does so by default). For a quick example of how to use the library, be sure to take a look in the test in `graph-processing/src/lib.rs`.
Closes #82