!8: Define pre-commit hooks to run before committing
The goal of this is to be able to do certain checks that are run in the CI before even creating a commit, allowing us to reduce the failure rate of pushed commits. This does not replace the CI, it is merely a quality-of-life addition. The pre-commit hooks are also not meant to be a 1:1 recreation of the CI, instead only running the simple checks.

For people using the nix devshell, everything is done for you automatically upon entering the shell. If you don't have (or want to use) nix dev environments, you can use these pre-commit hooks yourself by running `pre-commit install` after installing pre-commit with your package manager of choice. This will add the hooks to your local git repo.

This PR is more of a proposal and is open for discussion. If you think this should not be added then be sure to mention this.