!17: Feat/113/show streets and houses on the frontend
I made a working backend request in the branch "feat/98/making-working-backend-request", but there was some miscommunication and it wasn't properly pushed. I made this branch from the dev. This branch can be deleted after merge and the "feat/98/making-working-backend-request" should also be deleted without merging.

It has two working backend requests using proto and envoy proxy. The envoy proxy will later be revised as it is not secure yet. The backend requests can be found in the UI-backend cargo. The frontend code can be found in /frontend/src/app/dashboard/realtime.
 
The README.md in the frontend directory contains the commands to run the frontend and backend.