!21: Sensor data ingest service
This merge request implements a sensor data ingest service which can ingest a byte stream of data, turn it into the binary JSON (bson) format and store it in the archival database. This merge request also includes a library/binary for generating semi-realistic sensor data for use with the project.

For this merge request to be merged, the following has to be completed:
- [x] Integrate the sensor data generation with the ingest service
- [x] Decide on how to manage the database instances (how many postgres instances, tables vs databases, etc.)
- [x] Adding documentation to both public functions and lines of code.
- [x] Where to locate the migrations directory?
- [ ] ...

For the migrations directory, I propose to locate a single migrations directory in the root with a sub-directory per database. The running of these migrations should be done by ALL services that use the database.

Closes #33