!30: Add crate for performing database migrations at compile time
When building the project currently, you could get a lot of errors about tables not existing, not being able to connect to databases, etc. To simplify this process, this merge request introduces the ability to perform these migrations at compile time. This behavior togglable using the `COMPILE_TIME_MIGRATIONS` flag in the `.env` file. This behavior has to be implemented per service but is drastically simplified. An example can be found in sensor-data/ingestor/build.rs.

**IMPORTANT NOTE!**: Currently this crate does not check the shell environment for theses variables when they are not found in the `.env` file. I do not know if this would be wanted behavior.

In order to compile the project the only requirement is that you "upped" the docker containers (required databases are created).

This crate does introduce the need for tokio's async runtime in build scripts of services that use a database. It would be possible to avoid this by spawning a `sqlx-cli` shell process and waiting for it to finish. This would mean that error handling based on return values is no longer possible and errors would be worse in general (lack of context).