!35: Add `TODO.md` for small inconsistencies
This merge request adds a single [`TODO.md`](./TODO.md) to the root of the project. It contains a list of small inconsistencies/issues (not bugs) that do not belong to an issue and are too small to create an issue for. They can then be solved in a single merge request when someone has the time.