!65: Simulation optimizations
The simulation manager now perfoms setup and a full simulation run in parallel for the needed simulators. During the actual simulation, finished timeframes are sent to a database buffer which runs on a different thread. The last frame is also kept in memory to start the next step. This way the runner does not need to wait on the database every timestep.