!83: Feat/200/fix buildings selection
this functionality was broken by a merge in previous sprint. You should be able to select a building (building becomes red) you should be able to delete a building (building becomes grey) you should be able to revert delete of a (building becomes blue) buildings that are not selected are blue and you can only select one building at a time selection of transformation houses will be handled in next sprint(s).