!84: Fetch, create and delete sensors through the frontend
### Issue
Closes #203 


### Problem
The sensor operations on the frontend were only local and not reflected on the backend.


### Solution
This is now fixed by implementing an interface to fetch all sensors from the backend. This along with an api wrapper that is now used by the sensor page to create and delete sensors.


### Additional Notes
At the moment, when selecting quantities or units that are not supported by the backend, the user will be shown an error message stating this. This mismatch is planned (in this sprint or the next) to be solved by only defining quantities and units on the backend and telling the frontend which are supported.


### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [ ] Code has been tested and works (unit test)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline
- [x] Provided a clear title for MR 
- [x] Add your team's label 

## before review
- [x] Add Label ~"status:needs review"
- [x] Merge dev into task/203/sensor-frontend-connection
- [x] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue