!88: Add sensor data transformer to production docker compose setup
### Issue
Closes #206 


### Problem
The sensor data transformer was missing from the production docker compose setup. The production docker compose did not allow to run locally built images to test the production setup.


### Solution
The transformer is now added, use of an incorrect function in the generator is fixed and the docker compose now allows configuring the docker repository the images are pulled from (leave empty for local images). See `.env.example` for info on the environment variable.


### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase:
See changes :upside_down: 

**There appear to be a lot of extra changes, but this is because this MR builds on a list of previous (not merged) ones.**

### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [ ] Code has been tested and works (unit test)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline
- [x] Provided a clear title for MR 
- [x] Add your team's label 

## before review
- [x] Add Label ~"status:needs review"
- [ ] Merge dev into task/206/sensor-docker-compose
- [x] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue