!90: Request sensor signal quantities and units from backend
### Issue
Closes #208 


### Problem
Both the frontend and sensor store defined the quantities and units that are supported. This leads to mismatches and uncaught errors.


### Solution
This is solved by requesting supported units and quantities from the backend.


### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase:
I am not a frontend developer by any stretch of the imagination, thus would like feedback on how to improve this code.

### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [ ] Code has been tested and works (unit test)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline
- [x] Provided a clear title for MR 
- [x] Add your team's label 

## before review
- [x] Add Label ~"status:needs review"
- [x] Merge dev into task/208/quantity-unit-frontend-requests
- [x] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue