!97: feat/214/data-gen-new-dataset
### Issue
Provide a link to the original issue or bug report: #213 and #214 


### Problem
Describe the critical issue or bug being addressed: 
The previous dataset did contain weather data and energy consumption information from different geographical locations. The dataset also only contained information for one household. In order to make accurate predictions for energy consumption based on weather information, a larger dataset should be used. The sensor-data-generator should also be changed to adapt to the change of dataset.

### Solution
Explain the fix or solution implemented:
A new dataset was created using python pandas. This dataset aggregates weather information and energy consumption for multiple households situated in London. To clean and aggregate the data, pandas (python) was used. The sensor-data-generator now makes use of the new dataset.

### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase:
The sensor-data-generator was also converted to a cronjob. The interval and the timeframe should still be changed.


### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [x] Code has been tested and works (unit test)
- [x] Code has been formatted
- [x] No new issues introduced.
- [ ] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline
- [x] Provided a clear title for MR 
- [x] Add your team's label 

## before review
- [x] Add Label ~"status:needs review"
- [ ] Merge dev into feat/214/data-gen-new-dataset
- [x] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue