!120: (frontend/sensor) Fix first quantity in select field not being addable
### Issue
Provide a link to the original issue or bug report: /


### Problem
Describe the critical issue or bug being addressed: When first loading the signals page of the create sensor modal, the default quantity would show up correctly, but is not addable to this list of quantities. When selecting a different quantity and then going back to this one, it works again.


### Solution
Explain the fix or solution implemented: When fetching the quantities and units from the backend. Set the selected quantity as well.


### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase: /


### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made. (NA)
- [x] Code has been tested and works (unit test) (NA)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [x] Provided a clear title for MR 
- [x] Add your team's label 
- [x] Test using a local production setup

## before review
- [x] Add Label ~"status:needs review"
- [x] Merge dev into fix/first-quantity-not-selectable
- [x] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue