!131: Feat/251/userfriendly input in simulationmodal
This issue is more advanced to test:
When opening the editor or creating a simulation you should see the components on the right but instead of JSON format like previously, it is editable by the user. The data you see here is hardcoded in the frontend. Make sure the data is being saved by by clicking on another house and going back.

attention: when refreshing, the presets will reset as these are hardcoded (the solution is in another issue)

To test a more complex data structure please follow the following steps:
1. open frontend/src/components/maps/MapEditor.tsx
2. on line 52 you will find `const TypePresets = {` 
3. change the whole TypePresets object (from line 52 to line 83) to the contents of the following file:

[typePresets.txt](/uploads/c6cd83e80d16949012923a31c91aa423/typePresets.txt)


### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [ ] Code has been tested and works (unit test)
- [ ] Code has been formatted
- [ ] No new issues introduced.
- [ ] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [ ] Provided a clear title for MR 
- [ ] Add your team's label 
- [ ] Test using a local production setup

## before review
- [x] Add Label ~"status:needs review"
- [x] Merge dev into feat/251/userfriendly-input-in-simulationmodal
- [ ] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue