!136: !253: Unique sensor per building and unique quantity per sensor
### Issue

Provide a link to the original issue or bug report: Closes r2l/students/bacheloreindwerk2324/energie-simulator#253

### Problem

Describe the critical issue or bug being addressed:

When predicting sensor values, determining which sensor values to utilize for the forecast is not straightforward. However, by ensuring each quantity is unique per sensor, we ensure that there is always a singular sensor we need to address.

### Solution

Explain the fix or solution implemented:

Create a unique mapping between quantities, sensors and buildings.

### Additional Notes

Include any extra information or considerations for reviewers, such as impacted areas of the codebase:

### Merge Request Checklists

- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [ ] Code has been tested and works (unit test)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use `nix flake check -L` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [x] Provided a clear title for MR
- [x] Add your team's label
- [ ] Test using a local production setup

## before review

- [x] Add Label \~"status:needs review"
- [x] Merge dev into feat/253/unique-quantity-sensor-building
- [x] Request a review from developers that are familiar with the implementation

## after review

- [x] Remove Label \~"status:needs review" and add Label \~"status:reviewed"
- [x] Wait for at least 1 approval

## after merge request

- [x] Close your issue