!124: Prediction model
### Issue
Provide a link to the original issue or bug report: 
!198

### Solution
Explain the fix or solution implemented:
I implemented a vector auto regressor model so that the simulator can use it to predict future data.

### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase:
These is some math involved. If you have any questions about the matrices being used please go ahead and ask them.
There is an example in `crates/predictions/examples/`, run it using `cargo r -r --example vector_auto_regression`.
I'm also open and happy to discuss the performance of this implementation.

### Merge Request Checklists
- [ ] Code follows project coding guidelines.
- [ ] Documentation reflects the changes made.
- [ ] Code has been tested and works (unit test)
- [x] Code has been formatted
- [x] No new issues introduced.
- [ ] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [ ] Provided a clear title for MR 
- [x] Add your team's label 
- [ ] Test using a local production setup

## before review
- [ ] Add Label ~"status:needs review"
- [ ] Merge dev into feat/198/prediction-model
- [ ] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue