!127: Fix/check available global components before running simulations
### Issue
Provide a link to the original issue or bug report: #248 (part 1)


### Problem
Describe the critical issue or bug being addressed: The simulation was run even if not al needed global components were present. This resulted in crashing simulators.


### Solution
Explain the fix or solution implemented: Check the availability of the global components before starting the simulation.


### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase: use the following for the global components so that a simulation is actually run: {"global_time":{"unix_timestamp_millis":1000},"global_temperature":{"current_temp":15}}


### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [x] Code has been tested and works (unit test)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [x] Provided a clear title for MR 
- [x] Add your team's label 
- [x] Test using a local production setup

## before review
- [x] Add Label ~"status:needs review"
- [x] Merge dev into fix/check-available-components-for-simulations
- [x] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue