!146: (ui-backend/sensor): Add interface for (live) sensor data retrieval
### Issue
Provide a link to the original issue or bug report: #238 


### Problem
Describe the critical issue or bug being addressed: There was no interface to fetch ingested (transformed) sensor data.


### Solution
Explain the fix or solution implemented: There is now a service added to the ui-backend that allows the single use retrieval of sensor data based on a lookback. Most importantly, there is now an interface to fetch live sensor data from either all sensors or a single sensor. 


### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase: /


### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made. (NA)
- [x] Code has been tested and works (unit test) (NA)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [x] Provided a clear title for MR 
- [x] Add your team's label 
- [x] Test using a local production setup (NA)

## before review
- [x] Add Label ~"status:needs review"
- [x] Merge dev into feat/238/live-sensor-data-interface
- [x] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue