!137: Feat/256/display info on simulation status
### Issue
Provide a link to the original issue or bug report: closes #256 


### Problem
Describe the critical issue or bug being addressed: When simulations failed no info was provided on why the simulation was not able to run smoothly.


### Solution
Explain the fix or solution implemented: Store reason in database and provide to the frontend. Now when hovering over the simulation status symbol a short description appears.


### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase: to see info on the status hover over the symbol signaling the status of the simulation. To make a simulation you can reactivate one of the commented simulators in "docker-compose.prod.yml" or replace {"global_temperature":{"current_temp":15}} with {} on line 52 of "frontend/src/components/modals/CreateSimulationModal.tsx"


### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [x] Code has been tested and works (unit test)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [x] Provided a clear title for MR 
- [x] Add your team's label 
- [x] Test using a local production setup

## before review
- [x] Add Label ~"status:needs review"
- [x] Merge dev into feat/256/status-info
- [x] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue