!139: feat/194/choose-simulators-for-simulation
### Issue
Provide a link to the original issue or bug report: 


### Problem
Describe the critical issue or bug being addressed:
Creating the selection of simulators 

### Solution
Explain the fix or solution implemented:
Each simulators when they connect to the manager give an extra element which is its name. This then gets used in frontend to show the name of the simulators, for each one we add what components are changed by it. This is selected and than added to the createsimulationparams.

### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase:
Simulators, manager and frontend are changed

### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [x] Code has been tested and works (unit test)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [x] Provided a clear title for MR 
- [x] Add your team's label 
- [x] Test using a local production setup

## before review
- [x] Add Label ~"status:needs review"
- [x] Merge dev into feat/194/choose-simulators-for-simulation
- [x] Request a review from developers that are familiar with the implementation

## after review
- [x] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue