!144: Implement the realtime map page and connect it to the live sensor data interface
### Issue
Provide a link to the original issue or bug report: #230 


### Problem
Describe the critical issue or bug being addressed: The platform lacked a way to view sensor data that came in through a live interface.


### Solution
Explain the fix or solution implemented: This has been resolved on the realtime page where a heatmap is shown for sensor data that comes in.


### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase: This merge request also includes some (temporary) fixes. For example, mandatory `timestamp` signal on the frontend.


### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made. (NA)
- [x] Code has been tested and works (unit test) (NA)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [x] Provided a clear title for MR 
- [x] Add your team's label 
- [x] Test using a local production setup

## before review
- [x] Add Label ~"status:needs review"
- [x] Merge dev into feat/230/realtime-map
- [x] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue