!123: Feat/229/simulator for predictions
### Issue

Provide a link to the original issue or bug report: #229

### Problem

Describe the critical issue or bug being addressed:

A prediction model is used to predict energy consumption and weather related variables. Currently this prediction model is not used by any simulator.

### Solution

Explain the fix or solution implemented:

To use this prediction model in simulations, the supply-demand simulator will be changed so it uses the predicted values to simulate consumer and producer nodes. The temperature simulator will be replaced with a weather simulator that simulates temperature, precipitation, irradiance, wind speed and wind direction.

### Additional Notes

Include any extra information or considerations for reviewers, such as impacted areas of the codebase:

### Merge Request Checklists

- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [x] Code has been tested and works (unit test)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use `nix flake check -L` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [x] Provided a clear title for MR
- [x] Add your team's label
- [x] Test using a local production setup

## before review

- [x] Add Label \~"status:needs review"
- [x] Merge dev into feat/229/simulator-for-predictions
- [x] Request a review from developers that are familiar with the implementation

## after review

- [ ] Remove Label \~"status:needs review" and add Label \~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request

- [ ] Close your issue