!180: (realtime-sensors) Use bidirectional streams to allow the server to close connection threads
### Issue
Provide a link to the original issue or bug report: Closes #283


### Problem
Describe the critical issue or bug being addressed: Before, the server-side stream did not close properly when switching pages on the frontend. This would result in a resource munching thread that never stopped executing. This also meant that revisiting the realtime page would create more threads.


### Solution
Explain the fix or solution implemented: This used to be "solved" by always closing the server thread after x amount of time, but would result in the frontend losing connection with server after that same amount of time. This is now properly solved by switching to bidirectional streams that (on their own) would be able to shutdown the connection properly. This would, however, result in the tokio task not finishing (rust guarantees the correct cleanup of resources), thus not being able to display logs, send other signals, etc. This is solved by having the client send an explicit shutdown signal to the server, letting the servers task shut down cleanly and then close the connection.


### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase:


### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [x] Code has been tested and works (unit test) (NA)
- [x] Code has been formatted
- [x] No new issues introduced.
- [x] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [x] Provided a clear title for MR 
- [x] Add your team's label 
- [x] Test using a local production setup

## before review
- [x] Add Label ~"status:needs review"
- [x] Merge dev into fix/283/relax-realtime-sensor-data-stream
- [x] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue