!179: Fix simulation set to finished while still writing to the database
### Issue
closes #252 


### Problem
The simulation manager has a database buffer that writes simulation frames to the database on a different thread. The simulation status gets updated on the same thread as the simulation runner. If writing to the database takes much longer than running the simulation, the simulation will be marked as `Finished` while the buffer is still writing to the database.


### Solution
By also sending the status updates to the buffer, the simulation will only be marked as `Finished` after every frame has been written to the database.


### Additional Notes
Not every status update gets send to the buffer and will take effect immediately, as they happen before any frames get send to the buffer.


### Merge Request Checklists
- [x] Code follows project coding guidelines.
- [x] Documentation reflects the changes made.
- [ ] Code has been tested and works (unit test)
- [x] Code has been formatted
- [ ] No new issues introduced.
- [x] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [x] Provided a clear title for MR 
- [x] Add your team's label 
- [x] Test using a local production setup

## before review
- [x] Add Label ~"status:needs review"
- [x] Merge dev into fix/252/simulation-finished-when-not-ready
- [ ] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue