!176: remove code duplication in weather and energy simulator
### Issue

Provide a link to the original issue or bug report: #285

### Problem

Describe the critical issue or bug being addressed:

The energy-supply-and-demand simulator and the weather simulator required some similar functionality, which lead to code duplication.

### Solution

Explain the fix or solution implemented:

The common functionality was moved to a separate library called 'simulator-utilities'. I have chosen not to move the functionality to the sensor-store library since the functionality most likely will only be used for these specific simulators. This keeps the sensor-store library clean and easily understandable.

As an addition error handling was also improved.

### Additional Notes

Include any extra information or considerations for reviewers, such as impacted areas of the codebase:

### Merge Request Checklists

- [ ] Code follows project coding guidelines.
- [ ] Documentation reflects the changes made.
- [ ] Code has been tested and works (unit test)
- [ ] Code has been formatted
- [ ] No new issues introduced.
- [ ] If you have nix installed, you can use `nix flake check -L` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [ ] Provided a clear title for MR
- [ ] Add your team's label
- [ ] Test using a local production setup

## before review

- [ ] Add Label \~"status:needs review"
- [ ] Merge dev into feat/285/remove-code-dup-weather-energy-simu
- [ ] Request a review from developers that are familiar with the implementation

## after review

- [ ] Remove Label \~"status:needs review" and add Label \~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request

- [ ] Close your issue

Closes #285