!205: Smaller amount of historical data.
### Issue
Provide a link to the original issue or bug report: 
#314 

### Problem
Describe the critical issue or bug being addressed:

There is a bug where after some time, but within a few hours, that there are no new values in the sensor_values table.

### Solution
Explain the fix or solution implemented:

We can't replicate this bug locally. I reduced the amount of data pushed so that we prolong the inevitable crash.

### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase:


### Merge Request Checklists
- [ ] Code follows project coding guidelines.
- [ ] Documentation reflects the changes made.
- [ ] Code has been tested and works (unit test)
- [ ] Code has been formatted
- [ ] No new issues introduced.
- [ ] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [ ] Provided a clear title for MR 
- [ ] Add your team's label 
- [ ] Test using a local production setup

## before review
- [ ] Add Label ~"status:needs review"
- [ ] Merge dev into fix/314/smaller-amount-historical-data
- [ ] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue