!193: Feat/279/TUTORIAL AND DOCS
Attention, this MR is waiting for all frontend changes to be finished otherwise neither the tutorial and docs page will correspond to the latest changes. You can test this MR already or wait until all changes on the frontend are merged into dev (after this it has to be tested again).

- the website icon should redirect you to the home page.
- check out the docs page
- click on Start Tutorial
- go to the dashboard and select the tutorial drop(up) in the sidebar and test ALL tutorials. 

There have been bugs where the tutorial suddenly starts in the middle even though you are not following a tutorial so all functionalities should be tested without tutorials either. (yes, it is a lot of work)