!199: Resolve "Create weather/energy-supply-and-demand simulator README files"
### Issue
Provide a link to the original issue or bug report: #311 


### Problem
Describe the critical issue or bug being addressed:

Currently, there is no documentation in the form of README files for the energy-supply-and-demand simulator and the weather simulator. Additionaly, information regarding the origin of the datasets used should be mentioned somewhere as well.
This merge request is pure documentation!


### Solution
Explain the fix or solution implemented: Create README files.


### Additional Notes
Include any extra information or considerations for reviewers, such as impacted areas of the codebase: (merge request should only contain documentation)


### Merge Request Checklists
- [ ] Code follows project coding guidelines.
- [ ] Documentation reflects the changes made.
- [ ] Code has been tested and works (unit test)
- [ ] Code has been formatted
- [ ] No new issues introduced.
- [ ] If you have nix installed, you can use ```nix flake check -L``` to see if it will pass the pipeline. If you don't have nix, use `tools/run-pipeline.sh`.
- [ ] Provided a clear title for MR 
- [ ] Add your team's label 
- [ ] Test using a local production setup

## before review
- [ ] Add Label ~"status:needs review"
- [ ] Merge dev into feat/311/create-weather-energy-supply-and-demand-simulator-readme-files
- [ ] Request a review from developers that are familiar with the implementation

## after review
- [ ] Remove Label ~"status:needs review" and add Label ~"status:reviewed"
- [ ] Wait for at least 1 approval

## after merge request
- [ ] Close your issue

Closes #311